Send map state population data as floats, rather than bytes

+++ Sanctuary.Census.Common/Objects/RealtimeEvents/MapState.cs
@@ -46,7 +46,7 @@ public record MapState
     int CtfFlags,
     int RemainingCtfFlags,
     ValueEqualityDictionary<FactionDefinition, int> FactionPopulationUpperBound,
-    ValueEqualityDictionary<FactionDefinition, byte> FactionPopulationPercentage
+    ValueEqualityDictionary<FactionDefinition, float> FactionPopulationPercentage
 ) : IRealtimeEvent, ISanctuaryCollection
 {
     /// <inheritdoc />
